+++ baybe/utils/chemistry.py
@@ -4,6 +4,8 @@ import os
 import ssl
 import tempfile
 import urllib.request
+import warnings
+from collections.abc import Sequence
 from functools import lru_cache
 from pathlib import Path
 
@@ -12,16 +14,14 @@ import pandas as pd
 from joblib import Memory
 
 from baybe._optional.chem import (
-    Calculator,
+    BaseFingerprintTransformer,
     Chem,
-    GetMorganFingerprintAsBitVect,
-    RDLogger,
-    descriptors,
+    ConformerGenerator,
+    MolFromSmilesTransformer,
+    fingerprints,
 )
+from baybe.parameters.enum import SubstanceEncoding
 from baybe.utils.numerical import DTypeFloatNumpy
-
-_mordred_calculator = Calculator(descriptors)
-
 
 # Caching
 _cachedir = os.environ.get(
@@ -33,7 +33,9 @@ def _dummy_wrapper(func):
     return func
 
 
-_disk_cache = _dummy_wrapper if _cachedir == "" else Memory(Path(_cachedir)).cache
+_disk_cache = (
+    _dummy_wrapper if _cachedir == "" else Memory(Path(_cachedir), verbose=0).cache
+)
 
 
 def name_to_smiles(name: str) -> str:
@@ -72,158 +74,117 @@ def name_to_smiles(name: str) -> str:
 
 @lru_cache(maxsize=None)
 @_disk_cache
-def _smiles_to_mordred_features(smiles: str) -> np.ndarray:
-    """Memory- and disk-cached computation of Mordred descriptors.
+def _molecule_to_fingerprint_features(
+    molecule: str | Chem.Mol,
+    encoder: BaseFingerprintTransformer,
+) -> np.ndarray:
+    """Compute molecular fingerprint for a single molecule.
 
     Args:
-        smiles: SMILES string.
+        molecule: SMILES string or molecule object.
+        encoder: Instance of the fingerprint class to be used for computation.
 
     Returns:
-        Mordred descriptors for the given smiles string.
+        Array of fingerprint features.
     """
-    try:
-        return np.asarray(
-            _mordred_calculator(Chem.MolFromSmiles(smiles)).fill_missing()
-        )
-    except Exception:
-        return np.full(len(_mordred_calculator.descriptors), np.nan)
+    return encoder.transform([molecule])
 
 
-def smiles_to_mordred_features(
-    smiles_list: list[str],
-    prefix: str = "",
-    dropna: bool = True,
+def smiles_to_fingerprint_features(
+    smiles: Sequence[str],
+    encoding: SubstanceEncoding,
+    prefix: str | None = None,
+    kwargs_conformer: dict | None = None,
+    kwargs_fingerprint: dict | None = None,
 ) -> pd.DataFrame:
-    """Compute Mordred chemical descriptors for a list of SMILES strings.
+    """Compute molecular fingerprints for a list of SMILES strings.
 
     Args:
-        smiles_list: List of SMILES strings.
-        prefix: Name prefix for each descriptor
-            (e.g., nBase --> <prefix>_nBase).
-        dropna: If ``True``, drops columns that contain NaNs.
+        smiles: Sequence of SMILES strings.
+        encoding: Encoding used to transform SMILES to fingerprints.
+        prefix: Name prefix for each descriptor (e.g., nBase --> <prefix>_nBase).
+        kwargs_conformer: kwargs for conformer generator
+        kwargs_fingerprint: kwargs for fingerprint generator
 
     Returns:
-        Dataframe containing overlapping Mordred descriptors for each SMILES
-        string.
+        Dataframe containing fingerprints for each SMILES string.
     """
-    features = [_smiles_to_mordred_features(smiles) for smiles in smiles_list]
-    descriptor_names = list(_mordred_calculator.descriptors)
-    columns = [prefix + "MORDRED_" + str(name) for name in descriptor_names]
-    dataframe = pd.DataFrame(data=features, columns=columns, dtype=DTypeFloatNumpy)
+    kwargs_fingerprint = kwargs_fingerprint or {}
+    kwargs_conformer = kwargs_conformer or {}
 
-    if dropna:
-        dataframe = dataframe.dropna(axis=1)
+    if encoding is SubstanceEncoding.MORGAN_FP:
+        warnings.warn(
+            f"Substance encoding '{encoding.name}' is deprecated and will be disabled "
+            f"in a future version. Use '{SubstanceEncoding.ECFP.name}' "
+            f"with 'fp_size' 1024 and 'radius' 4 instead.",
+            DeprecationWarning,
+        )
+        encoding = SubstanceEncoding.ECFP
+        kwargs_fingerprint.update({"fp_size": 1024, "radius": 4})
 
-    return dataframe
+    elif encoding is SubstanceEncoding.RDKIT:
+        warnings.warn(
+            f"Substance encoding '{encoding.name}' is deprecated and will be disabled "
+            f"in a future version. Use '{SubstanceEncoding.RDKIT2DDESCRIPTORS.name}' "
+            f"instead.",
+            DeprecationWarning,
+        )
+        encoding = SubstanceEncoding.RDKIT2DDESCRIPTORS
+
+    fingerprint_cls = get_fingerprint_class(encoding)
+    fingerprint_encoder = fingerprint_cls(**kwargs_fingerprint)
+
+    if fingerprint_encoder.requires_conformers:
+        mol_list = ConformerGenerator(**kwargs_conformer).transform(
+            MolFromSmilesTransformer().transform(smiles)
+        )
+    else:
+        mol_list = smiles
+
+    features = np.concatenate(
+        [
+            _molecule_to_fingerprint_features(mol, fingerprint_encoder)
+            for mol in mol_list
+        ]
+    )
+    name = f"{encoding.name}_"
+    prefix = prefix + "_" if prefix else ""
+    col_names = [
+        prefix + name + f.split("fingerprint")[1]
+        for f in fingerprint_encoder.get_feature_names_out()
+    ]
+    df = pd.DataFrame(features, columns=col_names, dtype=DTypeFloatNumpy)
+
+    return df
 
 
-def smiles_to_molecules(smiles_list: list[str]) -> list[Chem.Mol]:
-    """Convert a given list of SMILES strings into corresponding Molecule objects.
+def get_fingerprint_class(encoding: SubstanceEncoding) -> BaseFingerprintTransformer:
+    """Retrieve the fingerprint class corresponding to a given encoding.
 
     Args:
-        smiles_list: List of SMILES strings.
-
-    Returns:
-        List of corresponding molecules.
+        encoding: A substance encoding.
 
     Raises:
-        ValueError: If the SMILES does not seem to be chemically valid.
-    """
-    mols = []
-    for smiles in smiles_list:
-        try:
-            mol = Chem.MolFromSmiles(smiles)
-            if mol is None:
-                raise ValueError()
-            mols.append(mol)
-        except Exception as ex:
-            raise ValueError(
-                f"The SMILES {smiles} does not seem to be chemically valid."
-            ) from ex
-    return mols
-
-
-def smiles_to_rdkit_features(
-    smiles_list: list[str], prefix: str = "", dropna: bool = True
-) -> pd.DataFrame:
-    """Compute RDKit chemical descriptors for a list of SMILES strings.
-
-    Args:
-        smiles_list: List of SMILES strings.
-        prefix: Name prefix for each descriptor (e.g., nBase --> <prefix>_nBase).
-        dropna: If ``True``, drops columns that contain NaNs.
+        ValueError: If no fingerprint class for the specified encoding is found.
 
     Returns:
-        Dataframe containing overlapping RDKit descriptors for each SMILES string.
+        The fingerprint class.
     """
-    mols = smiles_to_molecules(smiles_list)
+    # Exception case
+    if encoding is SubstanceEncoding.RDKITFINGERPRINT:
+        return fingerprints.RDKitFingerprint
 
-    res = []
-    for mol in mols:
-        desc = {
-            prefix + "RDKIT_" + dname: DTypeFloatNumpy(func(mol))
-            for dname, func in Chem.Descriptors.descList
-        }
-        res.append(desc)
-
-    df = pd.DataFrame(res)
-    if dropna:
-        df = df.dropna(axis=1)
-
-    return df
-
-
-def smiles_to_fp_features(
-    smiles_list: list[str],
-    prefix: str = "",
-    dtype: type[int] | type[float] = int,
-    radius: int = 4,
-    n_bits: int = 1024,
-) -> pd.DataFrame:
-    """Compute standard Morgan molecule fingerprints for a list of SMILES strings.
-
-    Args:
-        smiles_list: List of SMILES strings.
-        prefix: Name prefix for each descriptor (e.g., nBase --> <prefix>_nBase).
-        dtype: Specifies whether fingerprints will have int or float data type.
-        radius: Radius for the Morgan fingerprint.
-        n_bits:Number of bits for the Morgan fingerprint.
-
-    Returns:
-        Dataframe containing Morgan fingerprints for each SMILES string.
-    """
-    mols = smiles_to_molecules(smiles_list)
-
-    res = []
-    for mol in mols:
-        RDLogger.logger().setLevel(RDLogger.CRITICAL)
-
-        fingerp = GetMorganFingerprintAsBitVect(mol, radius, nBits=n_bits).ToBitString()
-        fingerp = map(int, fingerp)
-        fpvec = np.array(list(fingerp))
-        res.append(
-            {prefix + "FP_" + f"{k + 1}": dtype(bit) for k, bit in enumerate(fpvec)}
-        )
-
-    df = pd.DataFrame(res)
-
-    return df
-
-
-def is_valid_smiles(smiles: str) -> bool:
-    """Test if a SMILES string is valid according to RDKit.
-
-    Args:
-        smiles: SMILES string to be tested.
-
-    Returns:
-        ``True`` if the provided SMILES is valid, ``False`` else.
-    """
     try:
-        mol = Chem.MolFromSmiles(smiles)
-        return mol is not None
-    except Exception:
-        return False
+        cls_name = next(
+            name
+            for name in dir(fingerprints)
+            if (encoding.name + "Fingerprint").casefold() == name.casefold()
+        )
+    except StopIteration as e:
+        raise ValueError(
+            f"No fingerprint class exists for the specified encoding '{encoding.name}'."
+        ) from e
+    return getattr(fingerprints, cls_name)
 
 
 def get_canonical_smiles(smiles: str) -> str: